Fix segmentation fault error
A seg fault was thrown when an empty tree was being destroyed

+++ BinarySearchTree.h
@@ -138,7 +138,8 @@
         // récursive privée deleteSubTree(Node*)
         //
         ~BinarySearchTree() {
-            deleteSubTree(_root);
+            if (_root != nullptr)
+                deleteSubTree(_root);
         }
 
     private:
